Moved auth intefaces into auth.interfaces
