+++ src/services/auth/auth.service.ts
@@ -6,40 +6,8 @@ import request from 'request';
 import { v4 as uuidv4 } from 'uuid';
 import { User } from '../../shared/db/models/User';
 import { UserService } from '../user/user.service';
+import { TokenSet, SpotifyUserData } from './auth.interfaces';
 
-export interface TokenSet {
-  accessToken: string;
-  refreshToken: string;
-}
-
-export interface SpotifyUserDataExplicitContent {
-  filter_enabled: boolean;
-  filter_locked: boolean;
-}
-
-export interface SpotifyUserDataExternalUrls {
-  spotify: string;
-}
-
-export interface SpotifyUserDataFollowers {
-  href: string;
-  total: number;
-}
-
-export interface SpotifyUserData {
-  country: string;
-  display_name: string;
-  email: string;
-  explicit_content: SpotifyUserDataExplicitContent;
-  external_urls: SpotifyUserDataExternalUrls;
-  followers: SpotifyUserDataFollowers;
-  href: string;
-  id: string;
-  images: string[];
-  product: string;
-  type: string;
-  uri: string;
-}
 export class AuthService {
   userService = new UserService();
